Don't use recursion to diff objects
Diffing changes between nested objects is problematic,
because when used together with the {$set: {}} modifier,
Mongo would believe that the value of the key should be
set to the diffed object, which would swap out the object
in its entirity instead of just applying the diff.

The solution if not to use recursion for diffing any
nested objects, but instead simply to use equality
checking. If the value is identical, don't attach it to
the diff, if it's not attach the newest value in full.