Use recursion to diff nested documents
Both the $set and $unset diff functions now use recursion
to perform a "deep diff" of models before calling an update
to the Mongo collection.

This is done by using dot notation, where every property that
should be updated is an immediate descendant of the modifier,
but its name can contain dots, in which case it will be resolved
as a modifier of a nested document.