Maintain a single whitespace between elements in templates
Removing all whitespace in templates can have side effects for how
the content is presented when it comes to inline elements, eg.
```html
<p>My name is <em>Fredrik</em></p>
```
is compiled into
```html
<p>My name is<em>Fredrik</em></p>
```

Even though this does increase the filesize slightly, I think it
definitely is expected behavior for this kind of inline space to
be preserved all the way in to the browser.