Fix problem with multi-line comments breaking compileAngularTemplates step due to invalid JS generated for the template
The issue was introduced in a47ecc2. When a ng.html file contains multi-line comments Meteor failed on the console with error like "Unexpected token ILLEGAL" due to the generated JS that adds the html to the template cache was broken.
For example this could break your app:
```
<div>
<!--
  <div>
      Containing comments on multiple lines breaks the build
  </div>
-->
</div>
```
In order to properly fix that - the best future proof approach is to JSON.stringify the strings as this ensures any JS special character is properly escaped and turned into string literal.