***BREAKING CHANGE*** templateUrl values for *.ng.html files inside meteor packages have been changed.
angular-meteor now prefixes the package name when adding templates to $templateCache. The colon in the package name is replaced with an underscore and the package name is separated from the template name with an underscore. So, if you have a package named my-app:my-package with a template at packages/my-app:my-template/client/views/my-template.ng.html, the templateUrl you use in the route should be:

  my-app_my-package_client/views/my-template.ng.html