Escape {{ }} in camera example
I am not sure how to actually escape {{party.imageData}}. It is in the source code, but as the code is being processed, {{party.imageData}} does not appear in the (camera API docs)[http://angular-meteor.com/api/camera]. Please propose fix. I noticed it because the example did not work for me. Then I looked it up in the video of Urigo and saw the curly braces.