Fix idStringify for recent versions of Meteor
Since commit meteor/meteor@b3096e93, MongoId and the _idStringify
(or _idParse) method is in a separate package.

diffArray currently assumes the _idStringify and _idParse methods to
be in LocalCollection, but this isn't true anymore in recent versions
of Meteor. Therefore, diffArray() throws an error on (for example)
Meteor 1.2.

This commit fixes this issue by using the mongo-id package if we cannot
find LocalCollection._idStringify and LocalCollection._idParse are
undefined.