Fix specs that would have broken Mongo
Old specs for changing a null field to an array expected `$set` to be an object
of the form `{'a.0':1, 'a.1':2, 'a.2':3}`. That format works for modifying an
array but throws an error when trying to modify `null`. The correct way is to
`$set` as `{'a': [1, 2, 3]}`.

The `lib` function `getUpdates` was already handling this correctly, but the
unit tests were reporting this correct behavior as an error.