fix(https://github.com/Urigo/angular-meteor/issues/1577) run throttledDigest in a non-reactive manner.
It's possible that a digest loop from one autorun function, triggers another autorun function to be ran. Before this commit, the second autorun thought it was nested in the first autorun. This has as a side effect that when the first autorun re-runs, it will stop the second autorun from running. This commit runs the digest in a non-reactive manner, which means it temporarily resets the current active tracker computation. Therefore no parent-child relation is created between the 2 autorun functions, solving this problem.